+++ test/small_eq_sum_real.jl
@@ -7,7 +7,7 @@
         @constraint(m, sum(weights .* x) == max_val)
         @objective(m, Max, max_val)
         optimize!(m)
-        com = JuMP.backend(m).optimizer.model.inner
+        com = CS.get_inner_model(m)
         @test is_solved(com)
         @test JuMP.termination_status(m) == MOI.OPTIMAL
         @test JuMP.objective_value(m) == 4